Split Bevy impls across multiple files
