+++ bevy_proto_backend/src/impls/bevy_impls/animation.rs
@@ -1,0 +1,3 @@
+use bevy::app::App;
+
+pub(super) fn register(_app: &mut App) {}
